Update image.js
I found on testing that when returning a JSON string from a database that it wasn't also placing the saved image into the editor, on further investigation the 'data.file.url' object that should have been applied to the img tag being added into the image instance was being returned as 'undefined'. 

After checking to see that the data.file object existed, it turned out that it's being handled correctly for a JSON object and being created as an array instead of an object. As such it needs to be addressed as array 0. 

To do this I created a function to check if the data.file was an array or not and treat accordingly. That said it could be handled as '(!data.file[0])' and the returns reversed instead of '(typeof data.file[0]!=="undefined")'. But that's nitpicking for the sake of semantics.