Fixes an issue affecting multi-line Text blocks
Basically, if the user would create new lines in a Text block by pressing Shift + Enter, when hitting the backspace key with the caret at the first position of any line (except the first one), the block would be merged into the previous one instead of merging the lines. Also, just the content one the first line would be kept.